requirements.txt
crewai
openai
python-dotenv
requests
beautifulsoup4
tiktoken
python-dotenv